@@ -3,18 +3,18 @@
   "version": "0.0.0",
   "private": true,
   "dependencies": {
-    "history": "4.7.2",
-    "moment": "^2.24.0",
-    "react": "^16.1.1",
-    "react-bootstrap": "^0.31.0",
-    "react-date-picker": "^7.5.0",
-    "react-dom": "16.1.1",
-    "react-moment": "^0.9.2",
-    "react-redux": "5.0.7",
-    "react-router-dom": "4.2.2",
-    "redux": "4.0.0",
+    "history": "5.3.0",
+    "moment": "^2.29.4",
+    "react": "^18.2.0",
+    "react-bootstrap": "^2.4.0",
+    "react-date-picker": "^8.4.0",
+    "react-dom": "18.2.0",
+    "react-moment": "^1.1.2",
+    "react-redux": "8.0.2",
+    "react-router-dom": "6.3.0",
+    "redux": "4.2.0",
     "redux-logger": "3.0.6",
-    "redux-thunk": "2.2.0"
+    "redux-thunk": "2.4.1"
   },
   "scripts": {
     "start": "react-scripts start",
